move local storage logic into module
